backend_api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
minio==7.2.15
azure-keyvault-secrets==4.9.0
azure-keyvault-keys==4.9.0
azure-identity==1.19.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==1.9.0
python-magic==0.4.24